skills-db-project/skills-report-micro-app/requirements.txt
Flask==2.2.2
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.2
google-cloud-storage==2.7.0
gunicorn==20.1.0
openpyxl==3.0.10
pandas==1.5.2
mockito==1.4.0
psycopg2-binary==2.9.5
pytest==7.2.0
